// apps/cli/src/data-ingest/species/input-wiki/wiki-info-type.ts
export interface WikiSummary {
  title: string
  content: string
  contentUrls: {
    desktop: string
    mobile: string
  }
  thumbnailImage: string
  credit: string
  imageInfoUrl: string
  license: string
  licenseUrl: string
}

export interface WikiSummaryResponse {
  type: string
  title: string
  displaytitle: string
  namespace: {
    id: number
    text: string
  }
  wikibase_item: string
  titles: {
    canonical: string
    normalized: string
    display: string
  }
  pageid: number
  thumbnail: {
    source: string
    width: number
    height: number
  }
  originalimage: {
    source: string
    width: number
    height: number
  }
  lang: string
  dir: string
  revision: string
  tid: string
  timestamp: string
  description: string
  description_source: string
  content_urls: {
    desktop: {
      page: string
      revisions: string
      edit: string
      talk: string
    }
    mobile: {
      page: string
      revisions: string
      edit: string
      talk: string
    }
  }
  extract: string
  extract_html: string
}

export interface WikiMediaResponse {
  batchcomplete: string
  query: {
    pages: Record<string, {
      imageinfo: WikiMediaImageInfo[]
    }>
  }
}

export interface WikiMediaImageInfo {
  descriptionurl: string
  extmetadata: Record<'LicenseShortName' | 'Artist' | 'Copyrighted', {value: string, source: string, hidden?: string}> & { LicenseUrl?:
    {value: string, source: string, hidden?: string}
  }
}